pyproject.toml
[project]
name = "van-rally"
version = "0.1.0"
description = "Roadsurfer Van-Rally Routes Checker"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "pyyaml",
]

[tool.setuptools]
py-modules = [
    "main", 
    "api_utils", 
    "data_utils", 
    "output_handler",
    "translations",
    "config_utils", 
    "cache_utils",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-mock",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"
